Refactor code and functions
