+++ src/routes/v1/feed.route.js
@@ -1,9 +1,10 @@
 const express = require('express');
-const { jwtAuthenticate } = require('../../config/passport');
+
+const jwtAuthenticate = require('../../middlewares/auth');
 const { feedController } = require('../../controllers');
 
 const router = express.Router();
 
-router.get('/list', (_, __, next) => next(), jwtAuthenticate, feedController.get);
+router.get('/list', jwtAuthenticate, feedController.get);
 
 module.exports = router;
